requirements.txt
fastapi 
uvicorn 
python-multipart 
PyMuPDF
pillow